requirements.txt
google-adk
python-dotenv
opencv-python
easyocr